Move LogisticsProfiler out of core.lib and tag debug fence
