+++ common/src/main/java/com/logistics/power/cable/CableNetwork.java
@@ -1,6 +1,6 @@
 package com.logistics.power.cable;
 
-import com.logistics.core.lib.LogisticsProfiler;
+import com.logistics.core.LogisticsProfiler;
 import com.logistics.core.lib.energy.EnergyCapabilityLookup;
 import com.logistics.core.lib.energy.IEnergyStorage;
 import com.logistics.core.lib.power.AbstractEngineBlockEntity;
